requirements.txt
fastapi
uvicorn
python-multipart
firebase-admin
scikit-learn
joblib
opencv-python-headless
pillow
numpy
requests